+++ apps/web/src/persistence/browser/client/index.ts
@@ -1,0 +1,1 @@
+export * from "@persistence/browser/client/save-user";
