+++ apps/web/src/persistence/browser/client/save-user.ts
@@ -1,5 +1,5 @@
 import type { User } from "@type/user";
-import { saveCookie } from "@persistence/client/cookie";
+import { saveCookie } from "@persistence/browser/client/cookie";
 
 export const saveUser = async (user: User) => {
   saveCookie({ key: "user", value: JSON.stringify(user) });
