--- apps/web/src/persistence/client/index.ts
@@ -1,1 +1,0 @@
-export * from "@persistence/client/save-user";
